pyproject.toml
[project]
name = "vault-semantic-search"
version = "0.1.0"
description = "Semantic search for your Obsidian vault using local embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9",
    "chromadb>=0.5",
    "rich>=13.0",
    "httpx>=0.27",
    "python-frontmatter>=1.1",
    "local-first-common",
]

[project.scripts]
vsearch = "vsearch.cli:app"

[tool.uv.sources]
local-first-common = {path = "../local-first-common", editable = true}

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vsearch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that require Ollama to be running",
]
